requirements.txt
anthropic>=0.49.0
websockets>=12.0
python-dotenv>=1.0.0
